JAMES-3214 Move MailRepositoryStore related classes to mailRepository-api
Without this interface it is impossible to generate MailRepository that the calling code will rely on.

mailRepository-api furthermore have many dependencies that are orthogonal to data-api and should not be included in it.
