+++ server/data/data-file/pom.xml
@@ -81,6 +81,10 @@
         <dependency>
             <groupId>${james.groupId}</groupId>
             <artifactId>james-server-mailrepository-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-mailrepository-api</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
